EPBDS-241 modifications that allow using JCR and PTR on a single JCR instance

+++ STUDIO/trunk/org.openl.rules.workspace.test/src/org/openl/rules/workspace/TestHelper.java
@@ -31,7 +31,13 @@ public class TestHelper {
         }
     }
 
+    public static void deleteTestFolder() {
+        deleteFile(new File(FOLDER_TEST));
+    }
+
     public static boolean clearDirectory(File dir) {
+        if (dir == null || !dir.isDirectory())
+            return false;
         for (File file: dir.listFiles()) {
             if (!deleteFile(file)) {
                 return false;
